Cleanup more unnecessary files
All of these aren't needed anymore after the build and bloat up the final flatpak a fair bit.

Also update the libhandy commit, since there has been a force push
to the branch. Ideally there should be a fork that tags the
versions Solanum uses so the commits don't disappear